test: add tests for differentiate_polynomial
- Created a dedicated test file `Tests/test_simple_diffrentiation_function.py`.
- Moved basic tests from `Tests/test_Calculus.py` to the new dedicated file.
- Added test coverage for floating-point values (using `math.isclose`), negative coefficients, zero coefficients, and mismatched array lengths.
- Cleaned up the `sys.path` append mechanism in `test_Calculus.py` to properly import math modules for the tests.
